session_data.get_url -> server_util.get_url
This function has nothing to do with SessionData!
