+++ lib/streamlit/web/bootstrap.py
@@ -25,7 +25,6 @@ from streamlit import config
 from streamlit import env_util
 from streamlit import net_util
 from streamlit import secrets
-from streamlit import session_data
 from streamlit import url_util
 from streamlit import util
 from streamlit import version
@@ -36,6 +35,7 @@ from streamlit.secrets import SECRETS_FILE_LOC
 from streamlit.source_util import invalidate_pages_cache
 from streamlit.watcher import report_watchdog_availability, watch_dir, watch_file
 from streamlit.web.server import Server, server_address_is_unix_socket
+from streamlit.web.server import server_util
 
 LOGGER = get_logger(__name__)
 
@@ -190,7 +190,7 @@ def _on_server_start(server: Server) -> None:
         else:
             addr = "localhost"
 
-        util.open_browser(session_data.get_url(addr))
+        util.open_browser(server_util.get_url(addr))
 
     # Schedule the browser to open using the IO Loop on the main thread, but
     # only if no other browser connects within 1s.
@@ -219,33 +219,33 @@ def _print_url(is_running_hello: bool) -> None:
 
     if config.is_manually_set("browser.serverAddress"):
         named_urls = [
-            ("URL", session_data.get_url(config.get_option("browser.serverAddress")))
+            ("URL", server_util.get_url(config.get_option("browser.serverAddress")))
         ]
 
     elif (
         config.is_manually_set("server.address") and not server_address_is_unix_socket()
     ):
         named_urls = [
-            ("URL", session_data.get_url(config.get_option("server.address"))),
+            ("URL", server_util.get_url(config.get_option("server.address"))),
         ]
 
     elif config.get_option("server.headless"):
         internal_ip = net_util.get_internal_ip()
         if internal_ip:
-            named_urls.append(("Network URL", session_data.get_url(internal_ip)))
+            named_urls.append(("Network URL", server_util.get_url(internal_ip)))
 
         external_ip = net_util.get_external_ip()
         if external_ip:
-            named_urls.append(("External URL", session_data.get_url(external_ip)))
+            named_urls.append(("External URL", server_util.get_url(external_ip)))
 
     else:
         named_urls = [
-            ("Local URL", session_data.get_url("localhost")),
+            ("Local URL", server_util.get_url("localhost")),
         ]
 
         internal_ip = net_util.get_internal_ip()
         if internal_ip:
-            named_urls.append(("Network URL", session_data.get_url(internal_ip)))
+            named_urls.append(("Network URL", server_util.get_url(internal_ip)))
 
     click.secho("")
     click.secho("  %s" % title_message, fg="blue", bold=True)
